+++ backend/shared/shared/common/validators.py
@@ -18,14 +18,10 @@ PHONE_NUMBER_REGEX = (
 # \d in Javascript matches [0-9] and has been replaced:
 POSTAL_CODE_REGEX = r"^[0-9]{5}$"
 
-# \w in Javascript matches [A-Za-z0-9_] and has been replaced:
-NAMES_REGEX = r"^[A-Za-z0-9_',.ÄÅÖäåö-][^\d!#$%&()*+/:;<=>?@[\\\]_{|}~¡¿÷ˆ]+$"
-
 # Please note that using a RegexValidator in a Django model field hardcodes the used
 # regular expression into the model and its migration but using a function does not.
 PHONE_NUMBER_REGEX_VALIDATOR = RegexValidator(PHONE_NUMBER_REGEX)
 POSTAL_CODE_REGEX_VALIDATOR = RegexValidator(POSTAL_CODE_REGEX)
-NAMES_REGEX_VALIDATOR = RegexValidator(NAMES_REGEX)
 
 
 def validate_phone_number(phone_number) -> None:
@@ -52,13 +48,15 @@ def validate_postcode(postcode) -> None:
 
 def validate_name(name) -> None:
     """
-    Function wrapper for NAMES_REGEX_VALIDATOR. If used as a validator in a
-    Django model's field this does not hardcode the underlying regular expression into
-    the migration nor into the model.
+    Validates name to be a non-empty string with no trailing or leading whitespace.
 
-    Raise ValidationError if the given value doesn't pass NAMES_REGEX_VALIDATOR.
+    Raise ValidationError if the given value is not a non-empty string with no trailing
+    or leading whitespace.
     """
-    NAMES_REGEX_VALIDATOR(name)
+    if not (isinstance(name, str) and name == name.strip() and name.strip()):
+        raise ValidationError(
+            "Name must be a non-empty string with no trailing or leading whitespace"
+        )
 
 
 def validate_json(value) -> None: