+++ frontend/benefit/applicant/src/components/applications/applicationList/ApplicationList.tsx
@@ -1,17 +1,7 @@
-import { Pagination, Select } from 'hds-react';
-import React from 'react';
-import LoadingSkeleton from 'react-loading-skeleton';
-import Container from 'shared/components/container/Container';
-import theme from 'shared/styles/theme';
+import ListContents from 'benefit/applicant/components/applications/applicationList/listItem/ListContents';
+import React, { PropsWithChildren } from 'react';
 import { OptionType } from 'shared/types/common';
 
-import {
-  $Heading,
-  $HeadingContainer,
-  $ListWrapper,
-  $OrderByContainer,
-  $PaginationContainer,
-} from './ApplicationList.sc';
 import ListItem from './listItem/ListItem';
 import useApplicationList from './useApplicationList';
 
@@ -19,108 +9,59 @@ export interface ApplicationListProps {
   heading: ((count: number) => React.ReactNode) | React.ReactNode;
   status: string[];
   isArchived?: boolean;
-  clientPaginated?: boolean;
-  itemsPerPage?: number;
-  initialPage?: number;
-  pageHref?: (index: number) => string;
   orderByOptions?: OptionType[];
   noItemsText?: React.ReactNode;
+  onListLengthChanged?: (isLoading: boolean, length: number) => void;
 }
 
-const ApplicationsList: React.FC<ApplicationListProps> = ({
+const ApplicationList: React.FC<PropsWithChildren<ApplicationListProps>> = ({
   heading,
   status,
   isArchived,
-  clientPaginated = false,
-  itemsPerPage = 25,
-  initialPage,
   orderByOptions,
-  noItemsText = '',
+  noItemsText,
+  children,
+  onListLengthChanged,
 }) => {
   const {
     list,
     shouldShowSkeleton,
     shouldHideList,
-    currentPage,
-    setPage,
     t,
     orderBy,
     setOrderBy,
-    language,
+    hasItems,
   } = useApplicationList({
     status,
     isArchived,
-    initialPage: clientPaginated ? initialPage : null,
     orderByOptions,
   });
 
-  if (shouldHideList && !noItemsText) return null;
-
-  const hasItems = list?.length > 0;
-  let items =
+  const items =
     list?.map((props) => <ListItem key={props.id} {...props} />) || [];
-  if (clientPaginated && !shouldShowSkeleton) {
-    items = items.slice(
-      currentPage * itemsPerPage,
-      (currentPage + 1) * itemsPerPage
-    );
-  }
 
   const headingText =
     heading instanceof Function ? heading(list.length) : heading;
 
   return (
-    <Container backgroundColor={theme.colors.silverLight}>
-      {shouldShowSkeleton && (
-        <>
-          <$HeadingContainer>
-            <LoadingSkeleton width="20%" />
-          </$HeadingContainer>
-          <$ListWrapper>
-            <ListItem isLoading />
-          </$ListWrapper>
-        </>
-      )}
-      {!shouldShowSkeleton && (
-        <>
-          <$HeadingContainer>
-            <$Heading>{headingText}</$Heading>
-            <$OrderByContainer>
-              {orderByOptions?.length > 1 && (
-                <Select<OptionType>
-                  id={`application-list-'${status.join('-')}-order-by`}
-                  options={orderByOptions}
-                  defaultValue={orderBy}
-                  onChange={setOrderBy}
-                  label={t('common:applications.list.common.sortOrder')}
-                  disabled={!hasItems}
-                />
-              )}
-            </$OrderByContainer>
-          </$HeadingContainer>
-          <$ListWrapper>
-            {items}
-            {!hasItems && noItemsText}
-          </$ListWrapper>
-          {hasItems && clientPaginated && (
-            <$PaginationContainer>
-              <Pagination
-                pageHref={() => '#'}
-                pageIndex={currentPage}
-                pageCount={Math.ceil(list.length / Math.max(1, itemsPerPage))}
-                paginationAriaLabel={t('common:utility.pagination')}
-                onChange={(e, index) => {
-                  e.preventDefault();
-                  setPage(index);
-                }}
-                language={language}
-              />
-            </$PaginationContainer>
-          )}
-        </>
-      )}
-    </Container>
+    <ListContents
+      list={list}
+      shouldShowSkeleton={shouldShowSkeleton}
+      shouldHideList={shouldHideList}
+      t={t}
+      orderBy={orderBy}
+      setOrderBy={setOrderBy}
+      orderByOptions={orderByOptions}
+      hasItems={hasItems}
+      headingText={headingText}
+      status={status}
+      items={items}
+      noItemsText={noItemsText}
+      onListLengthChanged={onListLengthChanged}
+    >
+      {children}
+    </ListContents>
   );
 };
 
-export default ApplicationsList;
+export default ApplicationList;